+++ AnalyticsMessageStatusStorageQueueInbloundAdapter/function.json
